+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -487,7 +487,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KYD71GMF31E1N4SW7DRS99BY"),
+  _id: z.string().default("01KYDMWD70517XDDMCJ9X5RPYP"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: CreateDatasourceStatus$inboundSchema,