+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -278,7 +278,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KYD71GMAZ4S6A7E74A2JJ5P4"),
+  _id: z.string().default("01KYDMWD6TC07CBJHHRP29S6JP"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: ListDatasourcesStatus$inboundSchema,