+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -187,7 +187,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KYD71GMDGV66Q3JX6MFPAW3M"),
+  _id: z.string().default("01KYDMWD6YMK4BGBF948G8NNZH"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: RetrieveDatasourceStatus$inboundSchema,