+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -221,7 +221,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KYD71GMKS34235B8W37QDBX8"),
+  _id: z.string().default("01KYDMWD74RWYDYFKJA5J3033F"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: UpdateDatasourceStatus$inboundSchema,