+++ packages/orq-rc/RELEASES.md
@@ -11149,3 +11149,13 @@ Based on:
 - [typescript v4.13.0-rc.4] packages/orq-rc
 ### Releases
 - [NPM v4.13.0-rc.4] https://www.npmjs.com/package/@orq-ai/node/v/4.13.0-rc.4 - packages/orq-rc
+
+## 2026-07-25 22:06:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.791.0 (2.924.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.13.0-rc.5] packages/orq-rc
+### Releases
+- [NPM v4.13.0-rc.5] https://www.npmjs.com/package/@orq-ai/node/v/4.13.0-rc.5 - packages/orq-rc